# api_doc.yaml
swagger: '2.0'
info:
  description: |
    API for Broadway Connected app
  version: 1.0.0
  title: Broadway Connected API
  termsOfService: http://swagger.io/terms/
  contact:
    email: apiteam@swagger.io
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
# host: petstore.swagger.io
# basePath: /v2
...
paths:
...
  /api/subscribe-to-tag:
    post:
      tags:
      - Notifications
      summary: Subscribe to tag
      operationId: subscribeToTag
      consumes:
      - application/json
      produces:
      - application/json
      parameters:
      - in: body
        name: Subscription data
        description: Subscription data
        schema:
          type: object
          required:
          - user_id
          - tag_id
          properties:
            user_id:
              type: integer
              format: int64
              example: '80'
            tag_id:
              type: integer
              format: int64
              example: '61'

      responses:
        100:
          description: You are already subscribed to this topic before
          schema:
            type: object
            properties:
              Message:
                type: string
                example: 'You are already subscribed to this topic before'
              Response:
                type: string
                example: '1000'
              UserBlocked:
                type: integer
                example: 0
              Result:
                type: object
                example: {}

        200:
          description: You has been subscribed successfully on this tag
          schema:
            type: object
            properties:
              Message:
                type: string
                example: 'You has been subscribed successfully on this tag'
              Response:
                type: string
                example: '2000'
              UserBlocked:
                type: integer
                example: 0
              Result:
                type: boolean
                example: true
...
host: virtserver.swaggerhub.com
basePath: /openGeeksLab/BroadwayConnected/1.0.0
schemes:
 - https
 - http